Favorite ViewModel
Base bir favori viewmodeli yazdım birkaç yerde aynı methodları kullanan viewmodel ve composable içerikleri vardı . Bundan dolayı gereksiz tanımlamalardan kaçındım . Şuan doğru bir şekilde anasayfa ve favori kısımları çalışıyor.
